updated the parser to also construct the AST of the SMT formula.
added utility functions for destroying and printing SMT formulas.

updated the test driver to print the formula AST after parsing.

updated the input test file to match the new syntax. look at the
comment at the top of smt_parser.y for a description of the input
format accepted by the parser.

updated Makefile to find include header files.

checked with valgrind on the simple test case. the only memory leak
seems to be from the lexer.



git-svn-id: https://svn.pls.sei.cmu.edu/pacc/projects/misc/ddd@1815 780269f3-a799-4dce-8944-02d22917db73