New options
  --qelim-sof 
  --qelim-sof-nosupport

SOF variable elimination heuristic. By default, variable counting is
based on the support of the diagram. With -nosupport, variable counting
is done on the DAG of the diagram.

  --dyn

Enables automatic dynamic variable reordering. This option is not safe 
with TDDs and quantification. Use with care.

  --syntactic
Uses syntactic version of a theory. Used to compare TDD with regular
BDDs. This option is not safe with many TDD-specific methods. Use with 
care. Best when combined with --dyn








git-svn-id: https://svn.pls.sei.cmu.edu/pacc/projects/misc/ddd@2227 780269f3-a799-4dce-8944-02d22917db73