Changed dynamic variable reordering to respect TDD constraints.
Not tested for correctness. Ran on various hard examples. No crashes.

At the end, the implementation is just a few lines of code. However,
it does not play well with CUDD. Current hack makes the library incompatible
with standard BDDs. 

This creates one problem with TDDs. In TDD code, we use regular BDDs
to represent cubes for quantification. Currently, such cubes cannot be
trusted after dynamic reordering. In the worst case, dynamic
reordering may remove nodes from a cube. This may have effect on
running time of qelim_solver, but not on precision.


git-svn-id: https://svn.pls.sei.cmu.edu/pacc/projects/misc/ddd@2231 780269f3-a799-4dce-8944-02d22917db73