[FIX] potential fix for lddBoxWiden2Recur bug
The case when widening of sub-trees creates constraints that are
higher- or same-level as the parent is not correctly handled.

The case is caused by

	  /* floor top constraint */
	  newVCons = THEORY->floor_cons (vCons);

At the moment, I do not remember why this floor operation was
necessary. It only makes sense for rationals. I believe this was to
make widening converge to an integer bound faster.

The current fix is a work-around for constructing the final LDD.
It is not clear whether this fixes the underlying bug or is just a
work around the assertion in LDD construction.